+++ tests/test_render.py
@@ -1,11 +1,23 @@
 # -*- coding: utf-8 -*-
 import werobot.reply
-import werobot.test
+import werobot.testing
 from werobot.utils import to_unicode
 
 
+def test_wechatreply_render():
+    message = werobot.testing.make_text_message('test')
+    reply = werobot.reply.WeChatReply(message, content='hello',
+                                      time=1359803261)
+    try:
+        reply.render()
+    except NotImplementedError:
+        pass
+    else:
+        raise
+
+
 def test_text_render():
-    message = werobot.test.make_text_message('test')
+    message = werobot.testing.make_text_message('test')
     reply = werobot.reply.TextReply(message, content='hello', time=1359803261)
     reply_message = """
     <xml>
@@ -21,16 +33,50 @@ def test_text_render():
     assert result == to_unicode(reply_message)
 
 
+def test_star():
+    message = werobot.testing.make_text_message('test')
+    reply = werobot.reply.TextReply(message, content='hello', time=1359803261,
+                                    star=True)
+    reply_message = """
+    <xml>
+    <ToUserName><![CDATA[test]]></ToUserName>
+    <FromUserName><![CDATA[test]]></FromUserName>
+    <CreateTime>1359803261</CreateTime>
+    <MsgType><![CDATA[text]]></MsgType>
+    <Content><![CDATA[hello]]></Content>
+    <FuncFlag>1</FuncFlag>
+    </xml>
+    """.strip().replace(" ", "").replace("\n", "")
+    result = reply.render().strip().replace(" ", "").replace("\n", "")
+    assert result == to_unicode(reply_message)
+
+
 def test_convert():
-    message = werobot.test.make_text_message('test')
+    message = werobot.testing.make_text_message('test')
     reply = werobot.reply.TextReply(message, content='中文', time=1359803261)
     assert reply.render()
-    reply = werobot.reply.TextReply(message, content=to_unicode('中文'), time=1359803261)
+    reply = werobot.reply.TextReply(message, content=to_unicode('中文'),
+                                    time=1359803261)
     assert reply.render()
+
+
+def test_articles():
+    message = werobot.testing.make_text_message('test')
+    reply = werobot.reply.ArticlesReply(message, content='hello',
+                                        time=1359803261)
+    for _ in range(10):
+        reply.add_article(0)
+
+    try:
+        reply.add_article(0)
+    except AttributeError:
+        pass
+    else:
+        raise
 
 
 def test_create_reply():
-    message = werobot.test.make_text_message('test')
+    message = werobot.testing.make_text_message('test')
     reply = werobot.reply.create_reply('hi', message)
     assert reply
     reply = werobot.reply.create_reply([
@@ -43,7 +89,7 @@ def test_create_reply():
         [
             "whtsky",
             "I wrote WeRoBot",
-            "https://secure.gravatar.com/avatar/0024710771815ef9b74881ab21ba4173?s=420",
+            "http://gravatar.com/avatar/0024710771815ef9b74881ab21ba4173",
             "http://whouz.com/"
         ]
     ], message)