+++ tests/test_server.py
@@ -1,3 +1,23 @@
+import hashlib
+import time
+
+from werobot import WeRoBot
+from werobot.utils import generate_token, py3k
+
 
 def test_signature_checker():
-    pass
+    token = generate_token()
+
+    robot = WeRoBot(token)
+
+    timestamp = str(int(time.time()))
+    nonce = '12345678'
+
+    sign = [token, timestamp, nonce]
+    sign.sort()
+    sign = ''.join(sign)
+    if py3k:
+        sign = sign.encode()
+    sign = hashlib.sha1(sign).hexdigest()
+
+    assert robot.check_signature(timestamp, nonce, sign)